+++ start/src/App.vue
@@ -1,5 +1,13 @@
-<script lang="ts" setup></script>
+<script lang="ts" setup>
+  import { ref } from 'vue'
 
-<template></template>
+  const value1 = ref(20)
+</script>
 
-<style lang="scss" scoped></style>
+<template>
+  <h1>{{ value1 }}</h1>
+
+  <f-slider v-model="value1" :step="2" background="blue" active-color="green" width="200px" disabled />
+</template>
+
+<style></style>